mcpagent/requirements.txt
# requirements.txt
google-adk>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
asyncio>=3.4.3
httpx>=0.25.0

# For MCP support
mcp>=0.1.0

# Optional: for enhanced logging and monitoring
structlog>=23.1.0
prometheus-client>=0.19.0